# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Intelligent AI Assistant API
  description: |
    Next-generation **Agentic AI Life Assistant** API designed to act as the user's conscious digital twin, 
    continuously learning, guiding, and optimizing real-world actions in real time.
    
    ## Features
    - **Voice Processing**: Speech-to-text and text-to-speech with Whisper and Coqui/ElevenLabs
    - **Calendar Integration**: Google Calendar synchronization with intelligent scheduling
    - **WhatsApp Messaging**: Business Cloud API integration for confirmations and updates
    - **Federated Learning**: Privacy-preserving machine learning with FedAvg and Differential Privacy
    - **Digital Avatar**: 2D/3D animated user representation during interactions
    - **Privacy Compliance**: GDPR, DIFC, and UAE Data Protection Law compliance
    
    ## Authentication
    All endpoints (except health checks and webhook verification) require JWT authentication.
    Include the Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - **Authentication endpoints**: 5 requests per minute
    - **Voice processing**: 20 requests per minute
    - **General API**: 100 requests per minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns structured error responses:
    ```json
    {
      "error": "Error description",
...
      "request_id": "uuid",
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /whatsapp/send:
    post:
      tags: [WhatsApp Integration]
      summary: Send WhatsApp message
      description: |
        Send a message via WhatsApp Business Cloud API.
        
        **Requirements:**
        - User must have opted in to WhatsApp notifications
        - Valid WhatsApp Business API configuration
        - Recipient must have WhatsApp account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient, content]
              properties:
                recipient:
                  type: string
                  description: Phone number in international format
                  example: "+1234567890"
                content:
                  type: string
                  maxLength: 4096
                  example: "Your meeting reminder: Team standup at 10 AM"
                message_type:
                  type: string
                  enum: [text, template, interactive]
                  default: text
                template_name:
                  type: string
                  description: Template name for template messages
                  example: "meeting_reminder"
                template_params:
                  type: object
                  description: Parameters for template messages
                  additionalProperties: true
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhatsAppMessageResponse'
  /whatsapp/webhook:
    post:
      tags: [WhatsApp Integration]
      summary: WhatsApp webhook handler
      description: |
        Handle incoming WhatsApp webhook events.
        
        **Webhook Events:**
        - Message delivery status updates
        - Incoming user messages
        - User responses to confirmations
        
        **Security:**
        - Webhook signature verification required
        - HMAC-SHA256 with shared secret
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WhatsApp webhook payload
              additionalProperties: true
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  # Federated Learning Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **Token Format:**
        ```
        Authorization: Bearer <jwt_token>
        ```
        
        **Token Lifecycle:**
        - Access tokens expire after 1 hour
        - Refresh tokens expire after 30 days
        - Automatic token rotation supported
  schemas:
...
    WhatsAppMessageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        thread_id:
          type: string
          format: uuid
          example: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        message_id:
          type: string
          description: WhatsApp message ID
          example: "wamid.123456789"
        direction:
          type: string
          enum: [inbound, outbound]
          example: "outbound"
        content:
          type: string
          example: "Your meeting reminder: Team standup at 10 AM"
        message_type:
          type: string
          enum: [text, template, interactive, image, document]
          example: "text"
        status:
          type: string
          enum: [sent, delivered, read, failed]
          example: "delivered"
        sent_at:
          type: string
          format: date-time
          example: "2024-01-15T09:30:00Z"
...